feat h-log 콘텐츠 관련 commnuity domain 추가

+++ hobbyhobby-community/hobbyhobby-api/src/main/java/org/v1/dto/CommunityListResponse.java
@@ -5,7 +5,7 @@ import org.v1.model.Community;
 import java.util.List;
 
 public record CommunityListResponse(
-        Long communityId,
+        Integer communityId,
         String communityName,
         String imageUrl
 ) {
